# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Compilarr",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/library/{id}": {
      "get": {
        "tags": [
          "Library"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResource"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Library"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryResource"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/LibraryResource"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryResource"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/qualityprofile": {
      "get": {
        "tags": [
          "QualityProfile"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualityProfileResource"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "QualityProfile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityProfileResource"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/QualityProfileResource"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityProfileResource"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/qualityprofile/{id}": {
      "get": {
        "tags": [
          "QualityProfile"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityProfileResource"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "QualityProfile"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityProfileResource"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/QualityProfileResource"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityProfileResource"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "QualityProfile"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlbumPolicy": {
        "type": "integer"
      },
...
      "JsonElement": {},
      "LibraryLayout": {
        "type": "integer"
      },
      "LibraryResource": {
        "required": [
          "id",
          "name",
          "rootPath",
          "layout",
          "namingTemplate",
          "sidecarOptions",
          "albumPolicy",
          "minTracksPerRealAlbum",
          "plexSectionId",
          "isDefault"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "rootPath": {
            "type": "string"
          },
          "layout": {
            "$ref": "#/components/schemas/LibraryLayout"
          },
          "namingTemplate": {
            "type": "string"
          },
          "sidecarOptions": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "albumPolicy": {
            "$ref": "#/components/schemas/AlbumPolicy"
          },
          "minTracksPerRealAlbum": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "plexSectionId": {
            "type": [
              "null",
              "string"
            ]
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
      "QualityProfileItemResource": {
        "required": [
          "name",
          "qualities",
          "allowed"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "qualities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualityProfileQualityResource"
            }
          },
          "allowed": {
            "type": "boolean"
          }
        }
      },
      "QualityProfileQualityResource": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "QualityProfileResource": {
        "required": [
          "id",
          "name",
          "upgradeAllowed",
          "cutoff",
          "minScore",
          "durationToleranceMs",
          "items"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "upgradeAllowed": {
            "type": "boolean"
          },
          "cutoff": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "minScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "durationToleranceMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualityProfileItemResource"
            }
          }
        }
      },
...
    }
  },
...
}